Add "Book List" button to mobile navigation bar
If the user login as staff or admin, the "Catalog" button will change to "Book List" button.
If the user want to change it, the user can click the "Book List" button once more, it will swap to "Catalog" page.
This function is only work for admin user and staff user, student user will only see the "Catalog" page.